Refactor keybow
