+++ keybow/lua-config.h
@@ -8,21 +8,6 @@
 
 #pragma once
 
-int hid_output;
-int midi_output;
-int has_tick;
-unsigned long long tick_start;
-lua_State* L;
-
-unsigned short last_media_keys;
-unsigned short media_keys;
-unsigned short modifiers;
-unsigned short pressed_keys[14];
-
-unsigned short mouse_buttons;
-signed short mouse_x;
-signed short mouse_y;
-
 int initLUA();
 void luaTick(void);
 int luaHandleKey(unsigned short key_index, unsigned short state);
